requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-filter==0.13.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.3.3
gunicorn==19.4.5
Markdown==2.6.6
psycopg2==2.6.1
shortuuid==0.4.3
whitenoise==3.0
